[IMP] card_reader: implement partial authorization
This is required by Mercury. A partial payment happens automatically
when the balance of a card (eg. prepaid credit card) is too low to pay
for the entire order. Mercury then only authorizes what is left on the
card, allowing the customer to pay for the rest through other
means (eg. cash or another card).