[IMP] card_reader: implement enhanced duplicate logic functionality
This is required by Mercury. This uses functionality implemented by
Mercury to avoid accidental duplicate transactions. These can occur when
eg. a POS user swipes a card twice due to network connectivity issues or
when the Mercury response gets lost. This also attempts to deal with
connectivity issues between the POS and the Odoo server, by resending
the request after a certain time.