[IMP] tests: Add the possibility to tag tests
Purpose: When running tests, all the tests for the installed/updated
files are done. This commit adds a 'set_tags' decorator that can be used to tag
tests. Combined with a new 'test-tags' CLI option, it adds the ability
to filter which tests are executed. For example, @set_tags('slow') will
add a tag 'slow' to the test. The CLI option 'test-tags="slow"' will
only run tests tagged 'slow'.