[IMP] tests: use Chrome headless instead of phantomjs
Purpose: phantomjs was abandonned by the main author
see: https://groups.google.com/forum/#!topic/phantomjs/9aI5d-LDuNE

The aim of this commit is to replace phantomjs by Chrome headless.
It provides a unittest class that can be used the same way the
phantom js class was used.

selenium_run Parameters:
    url_path: the page to load
    ready: the js object that will return True when the code is ready
    js_code: the js test code to evaluate on that page
    login: the login to use a pre-authenticated session
    max_tries: the number of tries to check if the test ended

To evaluate that the javascript test has passed, the Chrome browser
is polled each second * max_tries. If the string 'ok' is found in the
browser console, the test is evaluated as successful.

Finally, the selenium_run method is not mandatory. The class can be
used to automate the test from the python code only.