[IMP] base: prevent deleting sensitive data
To prevent users to delete sensitive data, we added a 'nodelete'
attribute in the <data> tag of xml file. Once we want to delete some
records, the unlink() method will check in the ir_model_data table
if the record can be deleted or not regarding this new tag. The behavior
is similar to the 'noupdate' tag.