[REF] *: BS4, replace 'btn-xs' by 'btn-sm'
In a previous commit, we removed the use of all 'btn-sm' classes are we
used it everywhere for default-size buttons instead of customizing the
size of those default-size buttons directly.

This commit proves it was even more necessary as the 'btn-xs' class does
not exist anymore in BS4 and we so can use the 'btn-sm' class instead.