[IMP] web: propagate button context on save that preceeds it
Purpose: buttons on form view may hold custom context. This context is
generally used to control some of the business flow.

Currently when hitting an action button a call on saveRecord is done
before actually calling the action. It allows to ensure the record
is created or updated before running the action on it.

Context defined on action button is not propagated to the create or
write. This commit improves that behavior.