[FIX] tests: use psutil to check Chrome open port
When Chrome headless is started and the remote debugging port is used by
another process, Chrome emits an error message on stderr but does not
terminate. If the other process is another instance of Chrome headless
the test will interfere with the other one.

With this commit, psutil is used to wait chrome for listening on the
appropriate port.