[IMP] delivery: invoicing policy of delivery price
The delivery carriers get a new field to specify the invoicing
policy like the product's one. The two cases are :
  - Estimated cost: the delivery price is charded directly on
the sale order based on the estimated fee
  - Real cost: the delivery price is noted on the sale
order (a so line with quantity=0 and estimated price). Once the
delivery is validated, the delivery price is recomputed and its
quantity is set to 1.

The bases carriers (fixed_on_rule and base) always use Estimated cost
as their price should not be changed between the sale order creation
and delivery

Task : 1908654