[IMP] models: faster ensure_one
Unpacking the list is faster than measuring its len
to know if there is only one item in the list.

This is especially noticeable on big lists
(>100.000 items).

An empty list will raise not enough values to unpack
A list of more than 1 element will raise too many values to unpack