[WIP] mimetype sniffing in _binary_record_content
Binary fields are now stored as raw data in attachments rather than in
datas, meaning it's possible to have a "raw" binary field in
attachments (probably didn't work before? To check).

However this means mimetype sniffing doesn't work when setting the
value if the binary field's content is base64 encoded to start with —
which is the case for pretty much all images — as it's going to guess
the mimetype of base64 garbage so hopefully application/octet-stream,
possibly text/plain.

This just patches out _binary_record_content to redo mimetype
sniffing, but it's not a great solution.