[ADD] l10n_tests_config: add a new module to configure account tests
This module is intended for tests only. If installed, the database will
be modified. As its sequence is 10, this module will be installed very
early, before account and the l10n_xx itself.

This module detects the l10n_xx module to be installed and adapt the
country code of the company of the system user. That way, the coa of the
l10n_xx module will be loaded for the current company.

For example, an Odoo instance with this argument:
    -i l10_ch,account --test-tags l10n_generic

Will change the system user company country to "CH" and the tests tagged
as "l10n_generic" will be executed with the country correctly set on the
company. This imply to have some generic account tests.