INV. ADJUSTMENTS 10. Can update inv. line quantity
When inventory line theoretical quantity doesn't correspond anymore to
quantity in quant, marks the line as outdated.
An user can select and call action to update inventory lines.
So, theoretical quantity will be updated, but counted quantity will not.

When an inventory is validated, only the difference between theoretical
and counted quantity will be took to adjust quant.