[FIX] tests: increase HttpCase wait ready time warning
When an HttpCase test is performing a browser_js test, the first part is
to wait for some js code to ensure that the test is ready to start.
A warning was added to detect abnormal durations.

Altough 2sec is an abnormal duration, this commit increase this duration
to avoid runbot_merge disturbance during freeze.