[FIX] website_sale: handle free SO
When a sale order is free, no transaction are created, it causes the non
confirmation of the sale order and a redirection to the page
/shop/confirmation that display a transaction and lead to bad
displaying.

To fix this behaviour, we are calling 'action_confirm' on the orders
with amount = 0, and redirect to the portal view of this sale order.