[IMP] rating, project: store rating count and avg
The rating_count and rating_avg are not stored and
should be since the products, projects, etc displaying
the rating will do much more read than write on those fields.
Also there is a need to sort products by rating, for that the
rating_avg field needs to be stored.

Issue related to the field being now stored:
Computes that depends on the rating_ids are not triggered
because rating_ids has a domain. To make it work we need to manually
call modified on the models that inherit the rating.mixin

The field rating_last_value who is stored is now working

Tests have been added to ensure that we can manipulate
the ratings and that the compute will be triggered.