[IMP] crm: onchange of customer from quick create updates details in lead
Before commit:
When we create an opportunity manually from the Kanban view and associate
a customer to the opportunity, the "Initial Contact Information" doesn't
appear in the "Followup" tab on the lead so we have to reassign a partner
second time on the opportunity.

After commit:
When we create an opportunity manually from the Kanban view and associate
a customer to the opportunity, the "Initial Contact Information" will appear
in the "Followup" tab on lead.

Onchange fields are not available on quick create a view so the values do not
updated on record so added the invisible fields to make onchange works fabulously.

task-1974504