[IMP] crm: select 'linked with existing customer' customer related information updated on the 'followup' tab
Before update:
When we creating opportunity from lead and select option 'linked with existing
customer' Then in the 'followup' Section should't it fillup the partner
related information.

After commite:
When we creating opportunity from lead and select option 'linked with existing
customer' Then in the 'followup' Section should it fillup the partner related
information.

Vlaues are not available on covert opporatunity data method so the values do
not appear on 'followup' tab so added the required data to make a partner
information visible in FollowUp tab.

task-1974504