[FIX] mrp: bug fixing while user updates the quantity to produce on production.
Before this commit:
When user decrease or increase Quantity to Produce, it will generate an activity
for user but the quantity it shows is wrong,
If user decreases the quantity from 10 to 8, message will be shown as,
12 quantity ordered instead of 10.
same case while increasing the quantity.

After this commit:
When user updates the quantity to produce,
message on the activity about the changing quantities of products,
in production raw material is properly set.

task-1819526