[IMP] website_sale, *: avoid page reload on shop option change
* website

Before this commit, when the user was changing the size or sequence of a
product, the page was reloaded. The browser was also showing a strange
warning telling the user he was leaving the page and if the user clicked
'leave' the page was reloaded.

After this commit, when the user changes the sequence or size of a
product or the ppr, user changes will be reflected immediately without
reloading the page. These changes are permanent and once it is done,
can not be discarded with the discard button. To inform the user about
this behavior, a dialog is displayed.

task-1960780

Co-authored-by: JPR <jpr@openerp.com>
Co-authored-by: qsm-odoo <qsm@odoo.com>