[IMP] web_editor, website: allow to set background video to snippets
This commit adds support to set background videos to all the snippets
that already support setting background images. We've utilized the media
dialog snippet for the same.

Also, the snippet option 'Background Image' is renamed to 'Background'
because this option will now allow setting both background image and
video.

task-1920630