[IMP] l10n_it_edi: eased checking for Codice Ficale/VAT on invoice validation
Before this commit, when validating an e-invoice there are blocking errors if
VAT or Codice fiscale fields are empty for Buyers who are not Italian nationalist.

After this commit, if VAT or Codice Fiscale fields are empty for Buyers who are not
Italian nationalist, system will allow invoice validation.

Also, if Company Tax Representative of the seller is not having a VAT number,
which is mandatory, system won't allow to validate the invoice.

Task: 1971512
Closes: #34511