[IMP] crm_livechat: If there's no operator then visitor should receive a message
Purpose of this commit:

Case 1:
A livechat button should always be active whether there is an operator or not.

Case 2:
When there is no operator online, then show a message to the visitor that the
an operator is not online and takes an email from the visitor so through which,
lead can be created so the operator can see that lead with the messages from
visitor. and later on, send mail to the visitor on a provided email that a lead
has been created and reviewed by our team.

Case 3:
When a visitor sends messages to the operator but if an operator does not reply
even after 2 mins then a visitor will be asked with an email and then lead is
created with the messages that are sent to the operator before entering email
and then chat window will open and still, visitor can send the messages. So, all
these messages before and after an email entered are added into a lead. so,
the operator can review lead and also the visitor will get an automatic mail of
their demand.

task-2030381