[IMP] mrp: duplicate SN alert trigger sooner
On a manufacturing order, a check on serial numbers duplication is made only
when setting the stock move lines as done.

To avoid loosing too much time in the process, a first check is made when the
user choose the serial number (the finished product one or a component one) to be used.
A non-blocking warning is triggered in two cases:
 - A finished serial number already produced
 - A component serial number already consumed

Task : 2002133