[FIX] account, sale_timesheet: access rights when validating invoice
Still trying to solve the same problem as 49eb24a46771792bec454f6c4693c19d2b5b05b6
When validate an invoice, if the user is member of the timesheet user
group and the billing user group, it is forbidden to create an analytic
line without a project set. When the invoice is created, the billing user
should be allow to 'read' the generated analytic lines.