[IMP] eLearning: adds an animation when leveling up
There is now an animation with the progress bar when the user
is levelling up. The progress goes back to 0 the bounds changes
and the motivational phrase changes with a fade out/fade in
animation.

If the user reaches the last rank,
the progress bar is simply set to 100%

PR : #34881
Task ID : 2032649