[REF] web: allow field width hardcoding
Changed the way some field types are displayed.

Before this commit, column width in a list were calculated according to
a given factor (integer), determining their weight.

Now, fields can be given a factor in pixels or in percentage to
force their width to be static.
The rest of the columns will still be calculated from their weight
and the remaining space.

Task 1985916