[IMP] website_sale_delivery: add more test on checkout delivery flow
Before this commit, it was not possible anymore to go through the ecommerce
checkout when `website_sale_delivery` module was installed.

Indeed, it would raise a traceback on the confirmation step if more than one
delivery carrier would be available (different qweb if more than one carrier).

Thus, it would not be possible to click on pay now and buy the product.

This commit improves the current test to ensure such flows are working.