[IMP] models: remove "select id where id" from _read
When the `select` returns nothing more than ids that are already known, there is
no need to make it at all.

This divide by 2 the number of queries of `_read` when it has to fetch only
fields that are stored in a different table (o2m, ...).

The test that is removed here worked previously not because `write` is raising
on deleted records but because it was a special case of `res.partner` where
`write` is overridden and does a `read` of a o2m afterwards.

Part of task-2061122