[IMP] point_of_sale: readonly pos.config when there is open session
Make the pos.config readonly in the front when there is an open
pos.session linked to it. This way, we can prevent users' frustration
when they try to edit a pos.config only to find out in the end that
they are not allowed to.

Of course the pos.config record can be modified again when the
pos.sessions linked to it are close.