[FIX] web: model-field-selector issue
While selecting an option from popover, widget will set that option value in
'field-selector' space and shows related fields list (if any) in popover. Now
close that popover and focus-in that field-selector will open pop-over and shows
list of field options but not only related to that set field-selector value, it
will show all the options. This is the issue.
It should be shows options only related to that  already set field-selector
value. Described issue has been fixed by this commit.

task - 2058702