[IMP] rating, *: add sorting and filtering by rating
* = website_rating, website_rating

It is now possible to filter and sort by rating, the widget is based on
the one from the product chatter. A reset function and some fixes have
been added to the widget.
Previously the comments without rating were taken into account to
compute the rating_avg. this is no longer the case.