[FIX] mail, test_mail: quote usernames with special characters in emails
A username might contain unicode characters, but also ', '"', or '\' characters.
These can be particularly troublesome if the name is handled naively.
E.g. take the name '\u1f427, pen pen' (perfectly legal name).
This would be encoded as:
"From: =?UTF-8?B?8J+QpywgcGVuIHBlbg==?= <pen@odoodoo.com>"
Note that the two tokens are clearly delimited, so there is no ambiguity
in segmenting the username.
If, however, we squash the decoded token into
"\u1f427, pen pen <pen@odoodoo.com>"
then, then ", " (COMMASPACE) being used as a delimiter means that getaddresses
is thrown off, and parses it as:
('', '\u1f427') and ('pen pen', '<pen@odoodoo.com>').
The name is incorrect in the second, and '\u1f980' is not a valid address.

When using such a username on Thunderbird, it is shown quoted when displayed;
however the quotes are not actually sent
(it is not necessary because the encoding removes any potential ambiguity).
Similarly, when getaddresses is called, it automatically removes the quotes.
Moreover, there is a quote method in the email module done exactly to handle
the strings that could cause issues when quoting.

opw 1923429