[IMP] tests: Allow msg with assertRaises
The unittest signature of `assertRaises` allows a named
parameter `msg`. (see the official documentation [1]).
The message is displayed in case of assertion failure.

The Odoo implementation of assertRaises does not implement it
but many occurences are found in Odoo tests.
Odoo developpers usually expect the same signature as found in
the unittest documentation.

This commit implements the `msg` parameter in a similar
way as unittest.

[1] https://docs.python.org/3/library/unittest.html#unittest.TestCase.assertRaises