[IMP] point_of_sale: simplify action button definition method
In the method define_action_button() support for options
`{after,before}` was broken due to the usage of an `index` variable...
that is not used afterward. Also the variable `i` that is actually used
is not updated accordingly and default to `undefined`...

So, knowing that the options `{after,before}` are currently not used in
our codebase, this commit removes their support resulting into a
drastically simpler implementation for this method.