[IMP] tests: log a warning when using browser_js during install
When an HttpCase starts to make requests using the browser_js method
during the install phase of Odoo, it happens that some required models
are not yet loaded from the database. When Odoo tries to render the UI,
some models could be missing.

It typically happens when using the '-u' on a module that use the
browser_js method in a test which is not tagged 'post_install'.

With this commit, a warning is logged if a test uses the browser_js
method without being post_install.