[IMP] tools: use cache on result of html_sanitize
When records inheriting `mail.thread` are created, there is a tracking message
that is posted on the chatter of each record, such as "partner created".

This message is the same for every record (it just depends on the model name)
however due to how `_create` is built the fields of type HTML will call
`html_sanitize` for every value, even if they are the same.

Since `html_sanitize` is rather costly in CPU, this commit caches its value with
a low cache size, so that successive calls with the same parameters will be
computed only once, all while keeping a fast lookup and low memory usage.

As an example, when creating 1000 partners the time spent in `html_sanitize` was
around 10% of the total creation time before the commit, and it is reduced to
insignificance after the commit (less than 0.1%).