[FIX] account: broken filter in reconciliation widget
From Accounting app under the kanban card "Bank" click on
"Reconcilile items". Select a line to reconcile with a selected partner
or counterpart (or input one) and then filter the underlying lines using
for example the name of the record

No filtering will occur, this is due to the domain used by the filtering
function which is matching all lines from the current partner instead of
restricting the search to the filtered elements.

The solution is to fix a logical condition in the domain only when the
partner_id is selected

opw-2070458