[FIX] account: removed unused accounting report
Before this commit:
when try to open or duplicate Aged Partner Balance report with the studio it
gives traceback.

As deprecated pdf reports are removed by this commit:
https://github.com/odoo/odoo/commit/e51b4f659fc5e4253d3156754ada21166c8d1bf8
'Aged Partner Balance' report is removed from manifest but was not removed
report. So, it is giving traceback.

After this commit:
'Aged Partner Balance' report is removed, so it will not be shown in the reports.

task-2070274